web/requirements3.6.txt
absl-py==1.4.0
amqp==5.2.0
astor==0.8.1
async-timeout==4.0.2
backports.weakref==1.0rc1
billiard==3.6.4.0
bleach==1.5.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
html5lib==0.9999999
imageio==2.16.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==3.4.4.19
packaging==21.3
Pillow==8.4.0
pkg-resources==0.0.0
prompt-toolkit==3.0.36
protobuf==3.6.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.1.1
redis==4.3.6
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tdqm==0.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.64.1
typing-extensions==4.1.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.16.0
zipp==3.6.0
